+++ actions/actions.go
@@ -37,14 +37,14 @@ import (
 	uuid "github.com/satori/go.uuid"
 
 	"github.com/fabric8-services/fabric8-wit/actions/rules"
+	"github.com/fabric8-services/fabric8-wit/actions/change"
 	"github.com/fabric8-services/fabric8-wit/application"
-	"github.com/fabric8-services/fabric8-wit/convert"
 )
 
 // ExecuteActionsByOldNew executes all actions given in the actionConfigList
 // using the mapped configuration strings and returns the new context entity.
 // It takes the old version and the new version of the context entity, comparing them.
-func ExecuteActionsByOldNew(ctx context.Context, db application.DB, userID uuid.UUID, oldContext convert.ChangeDetector, newContext convert.ChangeDetector, actionConfigList map[string]string) (convert.ChangeDetector, []convert.Change, error) {
+func ExecuteActionsByOldNew(ctx context.Context, db application.DB, userID uuid.UUID, oldContext change.Detector, newContext change.Detector, actionConfigList map[string]string) (change.Detector, []change.Change, error) {
 	if oldContext == nil || newContext == nil {
 		return nil, nil, errs.New("execute actions called with nil entities")
 	}
@@ -58,8 +58,8 @@ func ExecuteActionsByOldNew(ctx context.Context, db application.DB, userID uuid.
 // ExecuteActionsByChangeset executes all actions given in the actionConfigs
 // using the mapped configuration strings and returns the new context entity.
 // It takes a []Change that describes the differences between the old and the new context.
-func ExecuteActionsByChangeset(ctx context.Context, db application.DB, userID uuid.UUID, newContext convert.ChangeDetector, contextChanges []convert.Change, actionConfigs map[string]string) (convert.ChangeDetector, []convert.Change, error) {
-	var actionChanges []convert.Change
+func ExecuteActionsByChangeset(ctx context.Context, db application.DB, userID uuid.UUID, newContext change.Detector, contextChanges []change.Change, actionConfigs map[string]string) (change.Detector, []change.Change, error) {
+	var actionChanges []change.Change
 	for actionKey := range actionConfigs {
 		var err error
 		actionConfig := actionConfigs[actionKey]
@@ -93,7 +93,7 @@ func ExecuteActionsByChangeset(ctx context.Context, db application.DB, userID uu
 // executeAction executes the action given. The actionChanges contain the changes made by
 // prior action executions. The execution is expected to add/update their changes on this
 // change set.
-func executeAction(act rules.Action, configuration string, newContext convert.ChangeDetector, contextChanges []convert.Change, actionChanges *[]convert.Change) (convert.ChangeDetector, []convert.Change, error) {
+func executeAction(act rules.Action, configuration string, newContext change.Detector, contextChanges []change.Change, actionChanges *[]change.Change) (change.Detector, []change.Change, error) {
 	if act == nil {
 		return nil, nil, errs.New("rule can not be nil")
 	}
