+++ actions/rules/action_nil.go
@@ -1,7 +1,7 @@
 package rules
 
 import (
-	"github.com/fabric8-services/fabric8-wit/convert"
+	"github.com/fabric8-services/fabric8-wit/actions/change"
 )
 
 // ActionNil is a dummy action rule that does nothing and has no sideffects.
@@ -12,6 +12,6 @@ type ActionNil struct {
 var _ Action = ActionNil{}
 
 // OnChange executes the action rule.
-func (act ActionNil) OnChange(newContext convert.ChangeDetector, contextChanges []convert.Change, configuration string, actionChanges *[]convert.Change) (convert.ChangeDetector, []convert.Change, error) {
+func (act ActionNil) OnChange(newContext change.Detector, contextChanges []change.Change, configuration string, actionChanges *[]change.Change) (change.Detector, []change.Change, error) {
 	return newContext, nil, nil
 }
